conformance/extension/WEBCL_html_image/cl_commandQueue_enqueueWriteImage_with_HTMLImageElement.html

<!--
 Copyright (C) 2013 Samsung Electronics Corporation. All rights reserved.

 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided the following condition
 are met:

 1.  Redistributions of source code must retain the above copyright
     notice, this list of conditions and the following disclaimer.

 2.  Redistributions in binary form must reproduce the above copyright
     notice, this list of conditions and the following disclaimer in the
     documentation and/or other materials provided with the distribution.

 THIS SOFTWARE IS PROVIDED BY SAMSUNG ELECTRONICS CORPORATION AND ITS
 CONTRIBUTORS "AS IS", AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING
 BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
 FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL SAMSUNG
 ELECTRONICS CORPORATION OR ITS CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
 INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES(INCLUDING
 BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 DATA, OR PROFITS, OR BUSINESS INTERRUPTION), HOWEVER CAUSED AND ON ANY THEORY
 OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT(INCLUDING
 NEGLIGENCE OR OTHERWISE ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 -->
<html>
<head>
<meta charset="utf-8">
    <title>WebCLCommandQueue enqueueWriteImage with imageElement conformance test.</title>
    <link rel="stylesheet" href="../../../resources/js-test-style.css"/>
    <script src="../../../resources/js-test-pre.js"></script>
    <script src="../../../resources/cl_utils.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
try {
    var image = new Image();
    image.src = "../../../resources/pebbleHalf.jpg";
    image.onload = execute;
} catch(e) {
    testFailed("Unable to load HTML image element : " + e.name);
}
var webCLContext, webCLCommandQueue, webCLImageRead, webCLImageWrite, webCLImageReadWrite, origin, region, imageWidth, imageHeight, webCLImageDescriptorReadWrite, webCLImage;
var releasedImage, originInvalid, originOutOfBound, regionInvalid, regionOutOfBound;
var webCLContextNew, webCLCommandQueueNew, webCLImageNew, webCLEvent, webCLEventNew, releasedEvent;
function execute() {
    try {
        debug("WebCLCommandQueue enqueueWriteImage with imageElement test case. \n");
        debug("( WebCLImage image, CLboolean blockingWrite, sequence<CLuint> origin, sequence<CLuint> region, HTMLImageElement srcImage, optional sequence<WebCLEvent>? eventWaitList, optional WebCLEvent? event)\n");

        var wtu = WebCLTestUtils;
        webCLContext = wtu.createContext();
        webCLCommandQueue = wtu.createCommandQueue(webCLContext);
        origin = [0, 0];
        imageWidth = image.width;
        imageHeight = image.height;
        region = [imageWidth, imageHeight];
        webCLImageDescriptorReadWrite = wtu.getSupportedImageFormats(webCLContext, webcl.MEM_READ_WRITE, imageWidth, imageHeight);

        webCLImage = wtu.createImage(webCLContext, webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImage, true, origin, region, image);", "WEBCL_EXTENSION_NOT_ENABLED");
        if (!wtu.enableExtension(webcl, "WEBCL_html_sharing")) {
            throw { description : "Failed to enable extension : WEBCL_html_sharing. Exitting..." };
        }

        // Testing with image as MEM_READ_ONLY memory object.
        webCLImageRead = wtu.createImage(webCLContext, webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite);
        shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageRead, true, origin, region, image);");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageRead, false, origin, region, image);");

        // Testing with image as MEM_WRITE_ONLY memory object.
        webCLImageWrite = wtu.createImage(webCLContext, webcl.MEM_WRITE_ONLY, webCLImageDescriptorReadWrite);
        shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageWrite, true, origin, region, image);");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageWrite, false, origin, region, image);");

        // Testing with image as MEM_READ_WRITE memory object.
        webCLImageReadWrite = wtu.createImage(webCLContext, webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite);
        shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image);");
        shouldBeUndefined("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, false, origin, region, image);");

        // Testing for exceptions.
        releasedImage = wtu.createImage(webCLContext, webcl.MEM_READ_WRITE, webCLImageDescriptorReadWrite);
        wtu.release(releasedImage);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(releasedImage, true, origin, region, image);", "INVALID_MEM_OBJECT");
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, invalid_image);", "INVALID_HOST_PTR");

        originInvalid = [];
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, originInvalid, region, image);", "INVALID_VALUE");
        originInvalid = [0];
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, originInvalid, region, image);", "INVALID_VALUE");
        originOutOfBound = [imageWidth + 1, imageHeight + 1];
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, originOutOfBound, region, image);", "INVALID_VALUE");

        regionInvalid = [];
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, regionInvalid, image);", "INVALID_VALUE");
        regionInvalid = [0];
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, regionInvalid, image);", "INVALID_VALUE");
        regionOutOfBound = [imageWidth + 1, imageHeight + 1];
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, regionOutOfBound, image);", "INVALID_VALUE");

        webCLContextNew = wtu.createContext();
        webCLCommandQueueNew = wtu.createCommandQueue(webCLContextNew);
        webCLImageNew = wtu.createImage(webCLContext, webcl.MEM_READ_ONLY, webCLImageDescriptorReadWrite);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageNew, true, origin, region, image);", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image);", "INVALID_CONTEXT");

        webCLEvent = wtu.createEvent(webCLContext);
        webCLEventNew = wtu.createEvent(webCLContextNew);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image, [webCLEventNew], webCLEvent);", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image, [webCLEvent], webCLEventNew);", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueueNew.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image, [], webCLEvent);", "INVALID_CONTEXT");
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image, [], webCLEventNew);", "INVALID_CONTEXT");

        releasedEvent = wtu.createEvent(webCLContext);
        wtu.release(releasedEvent);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image, [releasedEvent], webCLEvent);", "INVALID_EVENT_WAIT_LIST");

        wtu.setStatus(webCLEvent, -1);
        shouldThrowExceptionName("webCLCommandQueue.enqueueWriteImage(webCLImageReadWrite, true, origin, region, image, [webCLEvent], webCLEventNew);", "EXEC_STATUS_ERROR_FOR_EVENTS_IN_WAIT_LIST");

         // FIXME To add test case to verify exception : INVALID_IMAGE_SIZE.

    } catch (e) {
        testFailed(e.description);
    }

    // Since execute() is called on loading of the image, it is necessary to insert js-test-post.js in script tag, within this function.
    var script = document.createElement('script');
    script.src = '../../../resources/js-test-post.js';
    script.type = 'text/javascript';

    script.onload = function () {};

    document.getElementsByTagName('head')[0].appendChild(script);
}
successfullyParsed = true;
</script>
</body>
</html>
